Split up files into multiple
