+++ World.py
@@ -1,0 +1,274 @@
+from CollectionState import CollectionState
+from Region import Region
+from Entrance import Entrance
+from Location import Location
+from RegionList import create_regions
+from DungeonList import create_dungeons
+from Rules import set_rules, set_shop_rules
+from Item import Item
+import logging
+import copy
+
+
+class World(object):
+
+    def __init__(self, settings):
+        self.shuffle = 'vanilla'
+        self.dungeons = []
+        self.regions = []
+        self.itempool = []
+        self.state = CollectionState(self)
+        self._cached_locations = None
+        self._entrance_cache = {}
+        self._region_cache = {}
+        self._location_cache = {}
+        self.required_locations = []
+        self.shop_prices = {}
+
+        # dump settings directly into world's namespace
+        # this gives the world an attribute for every setting listed in Settings.py
+        self.settings = settings
+        self.__dict__.update(settings.__dict__)
+        # rename a few attributes...
+        self.keysanity = self.shuffle_smallkeys != 'dungeon'
+        self.check_beatable_only = not self.all_reachable
+        # trials that can be skipped will be decided later
+        self.skipped_trials = {
+            'Forest': False,
+            'Fire': False,
+            'Water': False,
+            'Spirit': False,
+            'Shadow': False,
+            'Light': False
+        }
+        # dungeon forms will be decided later
+        self.dungeon_mq = {
+            'DT': False,
+            'DC': False,
+            'JB': False,
+            'BW': False,
+            'IC': False,
+            'GTG': False,
+            'FoT': False,
+            'FiT': False,
+            'WT': False,
+            'SpT': False,
+            'ShT': False,
+            'GC': False
+        }
+
+        self.can_take_damage = True
+        self.keys_placed = False
+
+
+    def copy(self):
+        ret = World(self.settings)
+        ret.skipped_trials = copy.copy(self.skipped_trials)
+        ret.dungeon_mq = copy.copy(self.dungeon_mq)
+        ret.big_poe_count = copy.copy(self.big_poe_count)
+        ret.can_take_damage = self.can_take_damage
+        ret.shop_prices = copy.copy(self.shop_prices)
+        ret.keys_placed = self.keys_placed
+        ret.id = self.id
+        create_regions(ret)
+        create_dungeons(ret)
+        set_rules(ret)
+
+        # connect copied world
+        for region in self.regions:
+            copied_region = ret.get_region(region.name)
+            for entrance in region.entrances:
+                ret.get_entrance(entrance.name).connect(copied_region)
+
+        # fill locations
+        for location in self.get_locations():
+            if location.item is not None:
+                item = Item(location.item.name, location.item.advancement, location.item.priority, location.item.type)
+                item.world = location.item.world
+                ret.get_location(location.name).item = item
+                item.location = ret.get_location(location.name)
+                item.location.locked = location.locked
+
+        # copy remaining itempool. No item in itempool should have an assigned location
+        for item in self.itempool:
+            new_item = Item(item.name, item.advancement, item.priority, item.type)
+            new_item.world = item.world
+            ret.itempool.append(new_item)
+
+        # copy progress items in state
+        ret.state.prog_items = copy.copy(self.state.prog_items)
+
+        set_shop_rules(ret)
+
+        return ret
+
+
+    def initialize_regions(self):
+        for region in self.regions:
+            region.world = self
+            for location in region.locations:
+                location.world = self
+
+
+    def initialize_items(self):
+        for item in self.itempool:
+            item.world = self
+        for region in self.regions:
+            for location in region.locations:
+                if location.item != None:
+                    location.item.world = self
+        for item in [item for dungeon in self.dungeons for item in dungeon.all_items]:
+            item.world = self
+
+
+    def random_shop_prices(self):
+        shop_item_indexes = ['7', '5', '8', '6']
+        self.shop_prices = {}
+        for region in self.regions:
+            if self.shopsanity == 'random':
+                shop_item_count = random.randint(0,4)
+            else:
+                shop_item_count = int(self.shopsanity)
+
+            for location in region.locations:
+                if location.type == 'Shop':
+                    if location.name[-1:] in shop_item_indexes[:shop_item_count]:
+                        self.shop_prices[location.name] = int(random.betavariate(1.5, 2) * 60) * 5
+
+
+    def get_region(self, regionname):
+        if isinstance(regionname, Region):
+            return regionname
+        try:
+            return self._region_cache[regionname]
+        except KeyError:
+            for region in self.regions:
+                if region.name == regionname:
+                    self._region_cache[regionname] = region
+                    return region
+            raise RuntimeError('No such region %s' % regionname)
+
+
+    def get_entrance(self, entrance):
+        if isinstance(entrance, Entrance):
+            return entrance
+        try:
+            return self._entrance_cache[entrance]
+        except KeyError:
+            for region in self.regions:
+                for exit in region.exits:
+                    if exit.name == entrance:
+                        self._entrance_cache[entrance] = exit
+                        return exit
+            raise RuntimeError('No such entrance %s' % entrance)
+
+
+    def get_location(self, location):
+        if isinstance(location, Location):
+            return location
+        try:
+            return self._location_cache[location]
+        except KeyError:
+            for region in self.regions:
+                for r_location in region.locations:
+                    if r_location.name == location:
+                        self._location_cache[location] = r_location
+                        return r_location
+        raise RuntimeError('No such location %s' % location)
+
+
+    def get_items(self):
+        return [loc.item for loc in self.get_filled_locations()] + self.itempool
+
+
+    # get a list of items that should stay in their proper dungeon
+    def get_restricted_dungeon_items(self):
+        itempool = []
+        if self.shuffle_mapcompass == 'dungeon':
+            itempool.extend([item for dungeon in self.dungeons for item in dungeon.dungeon_items])
+        if self.shuffle_smallkeys == 'dungeon':
+            itempool.extend([item for dungeon in self.dungeons for item in dungeon.small_keys])
+        if self.shuffle_bosskeys == 'dungeon':
+            itempool.extend([item for dungeon in self.dungeons for item in dungeon.boss_key])
+
+        for item in itempool:
+            item.world = self
+        return itempool
+
+
+    # get a list of items that don't have to be in their proper dungeon
+    def get_unrestricted_dungeon_items(self):
+        itempool = []
+        if self.shuffle_mapcompass == 'keysanity':
+            itempool.extend([item for dungeon in self.dungeons for item in dungeon.dungeon_items])
+        if self.shuffle_smallkeys == 'keysanity':
+            itempool.extend([item for dungeon in self.dungeons for item in dungeon.small_keys])
+        if self.shuffle_bosskeys == 'keysanity':
+            itempool.extend([item for dungeon in self.dungeons for item in dungeon.boss_key])
+
+        for item in itempool:
+            item.world = self
+        return itempool
+
+
+    def find_items(self, item):
+        return [location for location in self.get_locations() if location.item is not None and location.item.name == item]
+
+
+    def push_item(self, location, item):
+        if not isinstance(location, Location):
+            location = self.get_location(location)
+
+        # This check should never be false normally, but is here as a sanity check
+        if location.can_fill(self.state, item, False):
+            location.item = item
+            item.location = location
+
+            logging.getLogger('').debug('Placed %s [World %d] at %s [World %d]', item, item.world.id if hasattr(item, 'world') else -1, location, location.world.id if hasattr(location, 'world') else -1)
+        else:
+            raise RuntimeError('Cannot assign item %s to location %s.' % (item, location))
+
+
+    def get_locations(self):
+        if self._cached_locations is None:
+            self._cached_locations = []
+            for region in self.regions:
+                self._cached_locations.extend(region.locations)
+        return self._cached_locations
+
+
+    def get_unfilled_locations(self):
+        return [location for location in self.get_locations() if location.item is None]
+
+
+    def get_filled_locations(self):
+        return [location for location in self.get_locations() if location.item is not None]
+
+
+    def get_reachable_locations(self, state=None):
+        if state is None:
+            state = self.state
+        return [location for location in self.get_locations() if state.can_reach(location)]
+
+
+    def get_placeable_locations(self, state=None):
+        if state is None:
+            state = self.state
+        return [location for location in self.get_locations() if location.item is None and state.can_reach(location)]
+
+
+    def unlocks_new_location(self, item):
+        temp_state = self.state.copy()
+        temp_state.clear_cached_unreachable()
+        temp_state.collect(item)
+
+        for location in self.get_unfilled_locations():
+            if temp_state.can_reach(location) and not self.state.can_reach(location):
+                return True
+
+        return False
+
+
+    def has_beaten_game(self, state):
+        return state.has('Triforce')
+
